requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.42.0
python-multipart==0.0.22
Jinja2==3.1.6
PyYAML==6.0.3
python-dotenv==1.2.2
pyinstaller==6.13.0
pywebview==6.0
httpx>=0.27.0

dashscope==1.25.15
langchain==1.2.14
langchain-community==0.4.1
langchain-core==1.2.25
langchain-chroma==1.1.0
langchain-text-splitters==1.1.1
langgraph==1.1.4
chromadb==1.5.5

opencc-python-reimplemented==0.1.7
pillow==12.2.0
playwright==1.58.0
requests==2.33.1
beautifulsoup4==4.14.3
pypdf>=5.0.0,<7.0.0
